Create replacingh2tagswithbrtags
Instead of having the <h2> tag, I made the header go into the <p> and got rid of the <h2> tag. Then to make it visually pleasing, I made a break after the original text that was in the <h2> tag so that it looks like a header, and then did the <small> tag to make the text after smaller for each div class.