Create AdvancedLayoutFloatInlineElement
<span> tags let you make items in the same tag display differently than the rest of the elements in the tag. 
block-level elements take up the entire page whereas inline elements exist within those lines and only take up as much of the screen as they need to. text-align may only align blocks, not spans because they are inline elements. To do this correctly, replace the text align with float, which is used for inline elements.