Create MobileFriendlyConditional
Line 57: 
for h1: shrank the top margin 
decreasing the whitespace with line height 
conditional so that if the layout is above 500 then it will format to this mobile friendly display. 
for h2: decreased font size
adjusted margin
adding space above and taking away space below the subheading. 
for div: 
adding 12px on left and right of divs 
cutting the margin in half to squeeze more content on the screen. 
for p: 
made font size 20px
gave p height 24 px 
for small: 
gave 16px